# Populate and use abbreviation cache when parsing DWARF